Replace or Remove a SaItem

replace_sa_item() replaces a SaItem of a SAProcessing and remove_sa_item() removes a SaItem from a SAProcessing

This functions clear a SA-Processing by removing all the sa-item contained.

Usage

replace_sa_item(jsap, idx, jsa)

remove_sa_item(jsap, idx)

remove_all_sa_item(jsap)

transfer_series(jsap_from, jsap_to, selected_series, print_indications = TRUE)

Arguments

jsap

the SAProcessing to modify.

idx

index of the target SaItem.

jsa

the new SaItem.

jsap_from

The SA-Processing from which to take the series

jsap_to

The SA-Processing in which to paste the series

selected_series

The vector containing the series-to-update's names.

print_indications

A boolean to print indications on the processing status (optional)

Value

NULL returned invisibly

NULL returned invisibly

Details

If selected_series is missing, all series from jsap_from will be copied. In this context, the word series refers to sa-item.